order views

+++ amc/views/orders.py
@@ -2,20 +2,19 @@
 
 import json
 
-from flask import (Blueprint, views, render_template, request,
+from flask import (Blueprint, views, render_template,
                    redirect, url_for)
 
 from flask.ext.login import current_user, login_required
 
 from amc.models import (OrderModel, OrderProductModel,
-                        ShoppingTrolleyModel, TrolleyProductModel)
+                        OrderHistoryModel, PayModel)
+from amc.utils import now
 
 bp = Blueprint('order', __name__)
 
 
 class TrolleyView(views.MethodView):
-    """修改购物车信息如添加产品购买数量，
-    """
 
     template = 'front/shopping_trolley.html'
 
@@ -35,23 +34,43 @@ class TrolleyView(views.MethodView):
         return render_template(self.template)
 
 
-class TrolleyOrderView(views.MethodView):
+class TrolleyCommitView(views.MethodView):
+    """点击提交，生成订单后清空购物车
+       本来可以和TrolleyView放一起使用post方法来提交，
+       考虑到没有表单提交，所以分开用get请求提交订单"""
 
     @login_required
     def get(self):
-        # 点击提交，生成订单后清空购物车
         items = current_user.trolley.products
+        if not items:
+            return
 
-        # 创建订单，初始状态launch
-        if items:
-            order = OrderModel.create(user_id=current_user.id)
+        # 创建订单，载入历史
+        order = OrderModel.create(user_id=current_user.id)
+        OrderHistoryModel.create(
+            order_id=order.id,
+            status=order.status,
+            operator_id=current_user.id)
+
         # 填充订单信息，将购物车结账清算
         for item in items:
+            if not item.is_supplied:
+                # 某件产品库存不足
+                return
+
+            product = item.product
             OrderProductModel.create(
                 order_id=order.id,
                 product_id=item.product_id,
                 product_quantity=item.product_quantity,
-                product_price=item.product.price)
+                product_price=product.price)
+
+            # 修改库存
+            product.quantity -= item.product_quantity
+            product.date_updated = now()
+            product.save()
+
+            # 清空购物车
             item.delete()
 
         return redirect(url_for('user.order'))
@@ -69,106 +88,105 @@ class TrolleyItemsView(views.MethodView):
                 product = {}
                 product["product_id"] = item.product_id
                 product["quantity"] = item.product_quantity
-                product["price"] = item.product.price
+                product["price"] = item.product_price
                 product["name"] = item.product.name
                 products_list.append(product)
         return json.dumps(products_list)
 
 
-class TrolleyAddView(views.MethodView):
+class OrderCancelView(views.MethodView):
+    """用户取消订单"""
+
+    STATUS_ALLOW = [OrderModel.STATUS_LAUNCH, OrderModel.STATUS_CONFIRM]
 
     @login_required
-    def get(self):
-        product_id = int(request.args.get('product_id'))
-        product_quantity = int(request.args.get('product_quantity'))
-        trolley = current_user.trolley
-        if not trolley:
-            trolley = ShoppingTrolleyModel(
-                user_id=current_user.id)
-            trolley.save()
+    def get(self, id):
+        order = OrderModel.query.get(id)
+        if not order:
+            return
+        if order.user_id != current_user.id:
+            # 非当前用户的订单
+            return
+        if order.status not in self.STATUS_ALLOW:
+            # 订单处于不被允许取消的状态
+            return
 
-        for item in trolley.products:
-            if (product_id == item.product_id):
-                if (product_quantity == item.product_quantity):
-                    return json.dumps({"status": "exists"})
-                else:
-                    item.update(product_quantity=product_quantity)
-                    return json.dumps({"status":"updated"})
-        # 不存在该产品
-        trolley_product = TrolleyProductModel(
-            trolley_id=trolley.id,
-            product_id=product_id,
-            product_quantity=product_quantity)
-        trolley_product.save()
-        return json.dumps({"status":"created"})
+        # 修改库存
+        for item in order.products:
+            product = item.product
+            product.quantity += item.product_quantity
+            product.date_updated = now()
+            product.save()
+            # 这里不能delete，否则取消掉的订单为空订单
+            # item.delete()
 
+        # 更新状态，载入历史
+        order.update(
+            status=OrderModel.STATUS_CANCEL,
+            date_updated=now())
+        OrderHistoryModel.create(
+            order_id=id,
+            status=order.status,
+            operator_id=current_user.id)
 
-class TrolleyUpdateView(views.MethodView):
-
-    @login_required
-    def get(self):
-        product_id = int(request.args.get('product_id'))
-        product_quantity = int(request.args.get('product_quantity'))
-        trolley = current_user.trolley
-        if not trolley:
-            trolley = ShoppingTrolleyModel(
-                user_id=current_user.id)
-            trolley.save()
-
-        for item in trolley.products:
-            if (product_id == item.product_id):
-                item.update(product_quantity=product_quantity)
-        
-        products_list = []
-        products = trolley.products
-        for item in products:
-            product = {}
-            product["product_id"] = item.product_id
-            product["quantity"] = item.product_quantity
-            product["price"] = item.product.price
-            product["name"] = item.product.name
-            products_list.append(product)
-        return json.dumps(products_list)
+        return redirect(url_for('user.order'))
 
 
-class TrolleyDeleteView(views.MethodView):
+class OrderSuccessView(views.MethodView):
+    """用户确认收获，生成收款单"""
 
     @login_required
-    def get(self):
-        product_id = int(request.args.get('product_id'))
-        trolley = current_user.trolley
+    def get(self, id):
+        order = OrderModel.query.get(id)
+        if not order:
+            return
+        if order.user_id != current_user.id:
+            # 非当前用户的订单
+            return
+        if order.status != OrderModel.STATUS_DISPATCH:
+            # 订单处不处于货物发出状态
+            return
 
-        for item in trolley.products:
-            if (product_id == item.product_id):
-                item.delete()
-        
-        products_list = []
-        products = trolley.products
-        for item in products:
-            product = {}
-            product["product_id"] = item.product_id
-            product["quantity"] = item.product_quantity
-            product["price"] = item.product.price
-            product["name"] = item.product.name
-            products_list.append(product)
-        return json.dumps(products_list)
+        # 更新状态，载入历史
+        order.update(
+            status=OrderModel.STATUS_SUCCESS,
+            date_updated=now())
+        OrderHistoryModel.create(
+            order_id=id,
+            status=order.status,
+            operator_id=current_user.id)
+
+        # 生成收款单
+        PayModel.create(
+            order_id=id,
+            amount=order.order_price)
+
+        return redirect(url_for('user.order'))
+
+
+class OrderReturnView(views.MethodView):
+    """申请退货，暂时不处理"""
+
+    @login_required
+    def get(self, id):
+        pass
 
 
 bp.add_url_rule(
     '/trolley/',
     view_func=TrolleyView.as_view('trolley'))
 bp.add_url_rule(
-    '/trolley/order/',
-    view_func=TrolleyOrderView.as_view('order_trolley'))
+    '/trolley/commit/',
+    view_func=TrolleyCommitView.as_view('trolley_commit'))
 bp.add_url_rule(
     '/trolley/items/',
     view_func=TrolleyItemsView.as_view('items_trolley'))
 bp.add_url_rule(
-    '/trolley/add/',
-    view_func=TrolleyAddView.as_view('add_trolley'))
+    '/order/cancel/<int:id>/',
+    view_func=OrderCancelView.as_view('order_cancel'))
 bp.add_url_rule(
-    '/trolley/update/',
-    view_func=TrolleyUpdateView.as_view('update_trolley'))
+    '/order/success/<int:id>/',
+    view_func=OrderSuccessView.as_view('order_success'))
 bp.add_url_rule(
-    '/trolley/delete/',
-    view_func=TrolleyDeleteView.as_view('delete_trolley'))
+    '/order/return/<int:id>/',
+    view_func=OrderReturnView.as_view('order_return'))
